software/training/pyproject.toml
[project]
name = "training"
version = "0.1.0"
description = "Training hub for the Sorter: pull samples from Hive, train on Vast.ai, publish models back to Hive."
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "click>=8.1",
    "Pillow>=10.0",
    "numpy>=1.26",
    "ultralytics>=8.3",
]

[project.scripts]
train = "training.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/training"]

[tool.hatch.build.targets.sdist]
include = ["src/training", "README.md"]
